lex: use a buffered channel
This means that the code doesn't have to do a goroutine switch in between every
single token.  Leads a 30% speedup on my ledger file.  Further speedups probably
require switching away from the goroutine-based lexer, as pprof suggests that
the bulk of remaining time is spent in concurrency-related low level code.